move includes into project directory and create version header
