+++ test/source/greeter.cpp
@@ -1,5 +1,9 @@
 #include <doctest/doctest.h>
-#include <greeter.h>
+
+#include <greeter/greeter.h>
+#include <greeter/version.h>
+
+#include <string>
 
 TEST_CASE("Greeter") {
   using namespace greeter;
@@ -10,4 +14,9 @@ TEST_CASE("Greeter") {
   CHECK(greeter.greet(LanguageCode::DE) == "Hallo Tests!");
   CHECK(greeter.greet(LanguageCode::ES) == "¡Hola Tests!");
   CHECK(greeter.greet(LanguageCode::FR) == "Bonjour Tests!");
+}
+
+TEST_CASE("Greeter version") {
+  static_assert(std::string_view(GREETER_VERSION) == std::string_view("1.0"));
+  CHECK(std::string(GREETER_VERSION) == std::string("1.0"));
 }
